+++ pandapower/__init__.py
@@ -2,7 +2,7 @@ import os
 pp_dir = os.path.dirname(os.path.realpath(__file__))
 
 from pandapower._version import __version__, __format_version__
-from pandapower.auxiliary import *
+from pandapower.auxiliary import *  # TODO: some functions shouldn't be available with import pandapower
 from pandapower.std_types import *
 from pandapower.create import *
 from pandapower.convert_format import *
@@ -11,16 +11,28 @@ from pandapower.sql_io import to_postgresql, from_postgresql, delete_postgresql_
 from pandapower.powerflow import *
 from pandapower.optimal_powerflow import OPFNotConverged
 from pandapower.run import *
-from pandapower.opf import *
-from pandapower.toolbox_general_issues import *
-from pandapower.toolbox_info import *
-from pandapower.toolbox_elm_selection import *
-from pandapower.toolbox_data_modification import *
+from pandapower.toolbox import *  # to be removed -> in future via package namespace available
 from pandapower.groups import *
-from pandapower.toolbox_grid_modification import *
 from pandapower.diagnostic import *
 from pandapower.runpm import *
 from pandapower.pf.runpp_3ph import runpp_3ph
 
 import pandas as pd
 pd.options.mode.chained_assignment = None  # default='warn'
+
+# import pandapower packages
+import pandapower.control
+import pandapower.converter
+import pandapower.estimation
+import pandapower.grid_equivalents
+import pandapower.networks
+# import pandapower.opf  # no imports available yet
+# import pandapower.pf  # no imports available yet
+import pandapower.plotting
+# import pandapower.protection  # no imports available yet
+# import pandapower.pypower  # no imports available yet
+import pandapower.shortcircuit
+import pandapower.test
+import pandapower.timeseries
+import pandapower.toolbox
+import pandapower.topology
