+++ pandapower/networks/__init__.py
@@ -8,4 +8,3 @@ from pandapower.networks.mv_oberrhein import *
 from pandapower.networks.power_system_test_cases import *
 from pandapower.networks.simple_pandapower_test_networks import *
 from pandapower.networks.ieee_europen_lv_asymmetric import *
-
