+++ pandapower/test/loadflow/result_test_network_generator.py
@@ -6,7 +6,7 @@
 
 import pandapower as pp
 from pandapower.test.toolbox import add_grid_connection, create_test_line
-from pandapower.toolbox_general_issues import nets_equal
+from pandapower.toolbox.general_issues import nets_equal
 
 def result_test_network_generator2(net, sn_mva=1, skip_test_impedance=False):
     """ This is a generator for the result_test_network
