+++ pandapower/test/toolbox/grid_modification.py
@@ -3,407 +3,14 @@
 # Copyright (c) 2016-2023 by University of Kassel and Fraunhofer Institute for Energy Economics
 # and Energy System Technology (IEE), Kassel. All rights reserved.
 
-
 import copy
-
 import numpy as np
 import pandas as pd
 from pandas._testing import assert_series_equal
 import pytest
 
 import pandapower as pp
-import pandapower.toolbox_general_issues as tbgi
 import pandapower.networks as nw
-import pandapower.control
-from pandapower.test.toolbox import assert_net_equal
-
-
-def test_element_bus_tuples():
-    ebts = pp.element_bus_tuples()
-    assert isinstance(ebts, list)
-    assert len(ebts) >= 20
-    item = next(iter(ebts))
-    assert isinstance(item, tuple)
-    assert len(item) == 2
-    assert len({"line", "gen"} & {elm for (elm, buses) in ebts}) == 2
-    assert {buses for (elm, buses) in ebts} == {"bus", "to_bus", "from_bus", 'hv_bus', 'mv_bus',
-                                                'lv_bus'}
-    assert len(pp.element_bus_tuples(bus_elements=False, res_elements=True)) > \
-           1.5 * len(pp.element_bus_tuples(bus_elements=False, res_elements=False)) > 0
-
-
-def test_pp_elements():
-    elms = pp.pp_elements()
-    assert isinstance(elms, set)
-    assert "bus" in elms
-    assert "measurement" in elms
-    assert "sgen" in elms
-    assert len(pp.pp_elements(bus=False, other_elements=False, bus_elements=True,
-                              branch_elements=False)) == \
-           len(pp.element_bus_tuples(bus_elements=True, branch_elements=False))
-
-
-def test_branch_element_bus_dict():
-    bebd = pp.branch_element_bus_dict()
-    assert isinstance(bebd, dict)
-    assert len(bebd) >= 5
-    assert bebd["trafo"] == ["hv_bus", "lv_bus"]
-    bebd = pp.branch_element_bus_dict(include_switch=True)
-    assert "bus" in bebd["switch"]
-
-
-def test_count_elements():
-    case9_counts = {"bus": 9, "line": 9, "ext_grid": 1, "gen": 2, "load": 3}
-    net = nw.case9()
-    received = pp.count_elements(net)
-    assert isinstance(received, pd.Series)
-    assert received.to_dict() == case9_counts
-    assert pp.count_elements(net, bus=False).to_dict() == {
-        et: num for et, num in case9_counts.items() if et not in ["bus"]}
-    assert pp.count_elements(net, bus=False, branch_elements=False).to_dict() == {
-        et: num for et, num in case9_counts.items() if et not in ["bus", "line"]}
-    received = pp.count_elements(net, return_empties=True)
-    assert len(received.index) == len(pp.pp_elements())
-    assert set(received.index) == pp.pp_elements()
-
-
-def test_signing_system_value():
-    assert pp.signing_system_value("sgen") == -1
-    assert pp.signing_system_value("load") == 1
-    for bus_elm in pp.pp_elements(bus=False, branch_elements=False, other_elements=False):
-        assert pp.signing_system_value(bus_elm) in [1, -1]
-    try:
-        pp.signing_system_value("sdfjio")
-        assert False
-    except ValueError:
-        pass
-
-
-def test_opf_task():
-    net = pp.create_empty_network()
-    pp.create_buses(net, 6, [10, 10, 10, 0.4, 7, 7],
-                    min_vm_pu=[0.9, 0.9, 0.88, 0.9, np.nan, np.nan])
-    idx_ext_grid = 1
-    pp.create_ext_grid(net, 0, max_q_mvar=80, min_p_mw=0, index=idx_ext_grid)
-    pp.create_gen(net, 1, 10, min_q_mvar=-50, max_q_mvar=-10, min_p_mw=0, max_p_mw=60)
-    pp.create_gen(net, 2, 8)
-    pp.create_gen(net, 3, 5)
-    pp.create_load(net, 3, 120, max_p_mw=8)
-    pp.create_sgen(net, 1, 8, min_q_mvar=-50, max_q_mvar=-10, controllable=False)
-    pp.create_sgen(net, 2, 8)
-    pp.create_storage(net, 3, 2, 100, min_q_mvar=-10, max_q_mvar=-50, min_p_mw=0, max_p_mw=60,
-                      controllable=True)
-    pp.create_dcline(net, 4, 5, 0.3, 1e-4, 1e-2, 1.01, 1.02, min_q_from_mvar=-10,
-                     min_q_to_mvar=-10)
-    pp.create_line(net, 3, 4, 5, "122-AL1/20-ST1A 10.0", max_loading_percent=50)
-    pp.create_transformer(net, 2, 3, "0.25 MVA 10/0.4 kV")
-
-    # --- run and check opf_task()
-    out1 = pp.opf_task(net, keep=True)
-    assert out1["flexibilities_without_costs"] == "all"
-    assert sorted(out1["flexibilities"].keys()) == [i1 + i2 for i1 in ["P", "Q"] for i2 in [
-        "dcline", "ext_grid", "gen", "storage"]]
-    for key, df in out1["flexibilities"].items():
-        assert df.shape[0]
-        if "gen" in key:
-            assert df.shape[0] > 1
-    assert out1["flexibilities"]["Pext_grid"].loc[0, "index"] == [1]
-    assert np.isnan(out1["flexibilities"]["Pext_grid"].loc[0, "max"])
-    assert out1["flexibilities"]["Pext_grid"].loc[0, "min"] == 0
-    assert np.isnan(out1["flexibilities"]["Qext_grid"].loc[0, "min"])
-    assert out1["flexibilities"]["Qext_grid"].loc[0, "max"] == 80
-    assert sorted(out1["network_constraints"].keys()) == ["LOADINGline", "VMbus"]
-    assert out1["network_constraints"]["VMbus"].shape[0] == 3
-
-    # check delta_pq
-    net.gen.loc[0, "min_p_mw"] = net.gen.loc[0, "max_p_mw"] - 1e-5
-    out2 = pp.opf_task(net, delta_pq=1e-3, keep=True)
-    assert out2["flexibilities"]["Pgen"].shape[0] == 1
-
-    net.gen.loc[0, "min_p_mw"] = net.gen.loc[0, "max_p_mw"] - 1e-1
-    out1["flexibilities"]["Pgen"].loc[0, "min"] = out1["flexibilities"]["Pgen"].loc[
-                                                      0, "max"] - 1e-1
-    out3 = pp.opf_task(net, delta_pq=1e-3, keep=True)
-    for key in out3["flexibilities"]:
-        assert pp.dataframes_equal(out3["flexibilities"][key], out1["flexibilities"][key])
-
-    # check costs
-    pp.create_poly_cost(net, idx_ext_grid, "ext_grid", 2)
-    pp.create_poly_cost(net, 1, "gen", 1.7)
-    pp.create_poly_cost(net, 0, "dcline", 2)
-    pp.create_pwl_cost(net, 2, "gen", [[-1e9, 1, 3.1], [1, 1e9, 0.5]], power_type="q")
-    out4 = pp.opf_task(net)
-    for dict_key in ["flexibilities", "network_constraints"]:
-        for key in out4[dict_key]:
-            assert pp.dataframes_equal(out4[dict_key][key], out1[dict_key][key])
-    assert isinstance(out4["flexibilities_without_costs"], dict)
-    expected_elm_without_cost = ["gen", "storage"]
-    assert sorted(out4["flexibilities_without_costs"].keys()) == expected_elm_without_cost
-    for elm in expected_elm_without_cost:
-        assert len(out4["flexibilities_without_costs"][elm]) == 1
-
-
-def test_nets_equal():
-    tbgi.logger.setLevel(40)
-    original = nw.create_cigre_network_lv()
-    net = copy.deepcopy(original)
-
-    # should be equal
-    assert pp.nets_equal(original, net)
-    assert pp.nets_equal(net, original)
-
-    # detecting additional element
-    pp.create_bus(net, vn_kv=.4)
-    assert not pp.nets_equal(original, net)
-    assert not pp.nets_equal(net, original)
-    net = copy.deepcopy(original)
-
-    # detecting removed element
-    net["bus"].drop(net.bus.index[0], inplace=True)
-    assert not pp.nets_equal(original, net)
-    assert not pp.nets_equal(net, original)
-    net = copy.deepcopy(original)
-
-    # detecting alternated value
-    net["load"]["p_mw"][net["load"].index[0]] += 0.1
-    assert not pp.nets_equal(original, net)
-    assert not pp.nets_equal(net, original)
-    net = copy.deepcopy(original)
-
-    # detecting added column
-    net["load"]["new_col"] = 0.1
-    assert not pp.nets_equal(original, net)
-    assert not pp.nets_equal(net, original)
-    net = copy.deepcopy(original)
-
-    # not detecting alternated value if difference is beyond tolerance
-    net["load"]["p_mw"][net["load"].index[0]] += 0.0001
-    assert pp.nets_equal(original, net, atol=0.1)
-    assert pp.nets_equal(net, original, atol=0.1)
-
-    # check controllers
-    original.trafo.tap_side.fillna("hv", inplace=True)
-    net1 = original.deepcopy()
-    net2 = original.deepcopy()
-    pp.control.ContinuousTapControl(net1, 0, 1.0)
-    pp.control.ContinuousTapControl(net2, 0, 1.0)
-    c1 = net1.controller.at[0, "object"]
-    c2 = net2.controller.at[0, "object"]
-    assert c1 == c2
-    assert c1 is not c2
-    assert pp.nets_equal(net1, net2)
-    c1.vm_set_pu = 1.01
-    assert c1 != c2
-    assert pp.nets_equal(net1, net2, exclude_elms=["controller"])
-    assert not pp.nets_equal(net1, net2)
-
-
-def test_clear_result_tables():
-    net = nw.case9()
-    pp.runpp(net)
-    elms_to_check = ["bus", "line", "load"]
-    for elm in elms_to_check:
-        assert net["res_%s" % elm].shape[0]
-    pp.clear_result_tables(net)
-    for elm in elms_to_check:
-        assert not net["res_%s" % elm].shape[0]
-
-
-def test_add_column_from_node_to_elements():
-    net = nw.create_cigre_network_mv("pv_wind")
-    net.bus["subnet"] = ["subnet_%i" % i for i in range(net.bus.shape[0])]
-    net.sgen["subnet"] = "already_given"
-    net.switch["subnet"] = None
-    net_orig = copy.deepcopy(net)
-
-    branch_bus = ["from_bus", "lv_bus"]
-    pp.add_column_from_node_to_elements(net, "subnet", False, branch_bus=branch_bus)
-
-    def check_subnet_correctness(ntw, elements, branch_bus_el):
-        for elm in elements:
-            if "bus" in ntw[elm].columns:
-                assert all(pp.compare_arrays(ntw[elm]["subnet"].values,
-                                             np.array(["subnet_%i" % bus for bus in ntw[elm].bus])))
-            elif branch_bus_el[0] in ntw[elm].columns:
-                assert all(pp.compare_arrays(ntw[elm]["subnet"].values, np.array([
-                    "subnet_%i" % bus for bus in ntw[elm][branch_bus_el[0]]])))
-            elif branch_bus_el[1] in ntw[elm].columns:
-                assert all(pp.compare_arrays(ntw[elm]["subnet"].values, np.array([
-                    "subnet_%i" % bus for bus in ntw[elm][branch_bus_el[1]]])))
-
-    check_subnet_correctness(net, pp.pp_elements(bus=False) - {"sgen"}, branch_bus)
-
-    pp.add_column_from_node_to_elements(net_orig, "subnet", True, branch_bus=branch_bus)
-    check_subnet_correctness(net_orig, pp.pp_elements(bus=False), branch_bus)
-
-
-def test_add_column_from_element_to_elements():
-    net = nw.create_cigre_network_mv()
-    pp.create_measurement(net, "i", "trafo", 5, 3, 0, side="hv")
-    pp.create_measurement(net, "i", "line", 5, 3, 0, side="to")
-    pp.create_measurement(net, "p", "bus", 5, 3, 2)
-    assert net.measurement.name.isnull().all()
-    assert ~net.switch.name.isnull().all()
-    orig_switch_names = copy.deepcopy(net.switch.name.values)
-    expected_measurement_names = np.array([
-        net.trafo.name.loc[0], net.line.name.loc[0], net.bus.name.loc[2]])
-    expected_switch_names = np.append(
-        net.line.name.loc[net.switch.element.loc[net.switch.et == "l"]].values,
-        net.trafo.name.loc[net.switch.element.loc[net.switch.et == "t"]].values)
-
-    pp.add_column_from_element_to_elements(net, "name", False)
-    assert all(pp.compare_arrays(net.measurement.name.values, expected_measurement_names))
-    assert all(pp.compare_arrays(net.switch.name.values, orig_switch_names))
-
-    del net.measurement["name"]
-    pp.add_column_from_element_to_elements(net, "name", True)
-    assert all(pp.compare_arrays(net.measurement.name.values, expected_measurement_names))
-    assert all(pp.compare_arrays(net.switch.name.values, expected_switch_names))
-
-
-def test_reindex_buses():
-    net_orig = nw.example_simple()
-    net = nw.example_simple()
-
-    to_add = 5
-    new_bus_idxs = np.array(list(net.bus.index)) + to_add
-    bus_lookup = dict(zip(net["bus"].index.values, new_bus_idxs))
-    # a more complexe bus_lookup of course should also work, but this one is easy to check
-    pp.reindex_buses(net, bus_lookup)
-
-    for elm in net.keys():
-        if isinstance(net[elm], pd.DataFrame) and net[elm].shape[0]:
-            cols = pd.Series(net[elm].columns)
-            bus_cols = cols.loc[cols.str.contains("bus")]
-            if len(bus_cols):
-                for bus_col in bus_cols:
-                    assert all(net[elm][bus_col] == net_orig[elm][bus_col] + to_add)
-            if elm == "bus":
-                assert all(np.array(list(net[elm].index)) == np.array(list(
-                    net_orig[elm].index)) + to_add)
-
-
-def test_continuos_bus_numbering():
-    net = pp.create_empty_network()
-
-    bus0 = pp.create_bus(net, 0.4, index=12)
-    pp.create_load(net, bus0, p_mw=0.)
-    pp.create_load(net, bus0, p_mw=0.)
-    pp.create_load(net, bus0, p_mw=0.)
-    pp.create_load(net, bus0, p_mw=0.)
-
-    bus0 = pp.create_bus(net, 0.4, index=42)
-    pp.create_sgen(net, bus0, p_mw=0.)
-    pp.create_sgen(net, bus0, p_mw=0.)
-    pp.create_sgen(net, bus0, p_mw=0.)
-
-    bus0 = pp.create_bus(net, 0.4, index=543)
-    pp.create_shunt(net, bus0, 2, 1)
-    pp.create_shunt(net, bus0, 2, 1)
-    pp.create_shunt(net, bus0, 2, 1)
-
-    bus0 = pp.create_bus(net, 0.4, index=5675)
-    pp.create_ward(net, bus0, 2, 1, 1, 2)
-    pp.create_ward(net, bus0, 2, 1, 1, 2)
-    pp.create_ward(net, bus0, 2, 1, 1, 2)
-
-    pp.create_continuous_bus_index(net)
-
-    buses = net.bus.index
-    assert all(buses[i] <= buses[i + 1] for i in range(len(buses) - 1))  # is ordered
-    assert all(buses[i] + 1 == buses[i + 1] for i in range(len(buses) - 1))  # is consecutive
-    assert buses[0] == 0  # starts at zero
-
-    used_buses = []
-    for element in net.keys():
-        try:
-            used_buses.extend(net[element].bus.values)
-        except AttributeError:
-            try:
-                used_buses.extend(net[element].from_bus.values)
-                used_buses.extend(net[element].to_bus.values)
-            except AttributeError:
-                try:
-                    used_buses.extend(net[element].hv_bus.values)
-                    used_buses.extend(net[element].lv_bus.values)
-                except AttributeError:
-                    continue
-
-    # assert that no buses were used except the ones in net.bus
-    assert set(list(used_buses)) - set(list(net.bus.index.values)) == set()
-
-
-def test_reindex_elements():
-    net = nw.example_simple()
-
-    new_sw_idx = [569, 763, 502, 258, 169, 259, 348, 522]
-    pp.reindex_elements(net, "switch", new_sw_idx)
-    assert np.allclose(net.switch.index.values, new_sw_idx)
-
-    net2 = copy.deepcopy(net)
-
-    previous_idx = new_sw_idx[:3]
-    new_sw_idx = [2, 3, 4]
-    pp.reindex_elements(net, "switch", new_sw_idx, previous_idx)
-    assert np.allclose(net.switch.index.values[:3], new_sw_idx)
-
-    # using lookup
-    pp.reindex_elements(net2, "switch", lookup=dict(zip(previous_idx, new_sw_idx)))
-    assert_net_equal(net, net2)
-
-    pp.reindex_elements(net, "line", [77, 22], [2, 0])
-    assert np.allclose(net.line.index.values, [22, 1, 77, 3])
-    assert np.allclose(net.switch.element.iloc[[4, 5]], [77, 77])
-
-    old_idx = copy.deepcopy(net.bus.index.values)
-    pp.reindex_elements(net, "bus", old_idx + 2)
-    assert np.allclose(net.bus.index.values, old_idx + 2)
-
-    pp.reindex_elements(net, "bus", [400, 600], [4, 6])
-    assert 400 in net.bus.index
-    assert 600 in net.bus.index
-
-
-def test_continuous_element_numbering():
-    from pandapower.estimation.util import add_virtual_meas_from_loadflow
-    net = nw.example_multivoltage()
-
-    # Add noises to index with some large number
-    net.line.rename(index={4: 280}, inplace=True)
-    net.trafo.rename(index={0: 300}, inplace=True)
-    net.trafo.rename(index={1: 400}, inplace=True)
-    net.trafo3w.rename(index={0: 540}, inplace=True)
-
-    net.switch.loc[(net.switch.et == "l") & (net.switch.element == 4), "element"] = 280
-    net.switch.loc[(net.switch.et == "t") & (net.switch.element == 0), "element"] = 300
-    net.switch.loc[(net.switch.et == "t") & (net.switch.element == 1), "element"] = 400
-    pp.runpp(net)
-    add_virtual_meas_from_loadflow(net)
-    assert net.measurement["element"].max() == 540
-
-    pp.create_continuous_elements_index(net)
-    assert net.line.index.max() == net.line.shape[0] - 1
-    assert net.trafo.index.max() == net.trafo.shape[0] - 1
-    assert net.trafo3w.index.max() == net.trafo3w.shape[0] - 1
-    assert net.measurement["element"].max() == net.bus.shape[0] - 1
-
-
-def test_scaling_by_type():
-    net = pp.create_empty_network()
-
-    bus0 = pp.create_bus(net, 0.4)
-    pp.create_load(net, bus0, p_mw=0., type="Household")
-    pp.create_sgen(net, bus0, p_mw=0., type="PV")
-
-    pp.set_scaling_by_type(net, {"Household": 42., "PV": 12})
-
-    assert net.load.at[0, "scaling"] == 42
-    assert net.sgen.at[0, "scaling"] == 12
-
-    pp.set_scaling_by_type(net, {"Household": 0, "PV": 0})
-
-    assert net.load.at[0, "scaling"] == 0
-    assert net.sgen.at[0, "scaling"] == 0
 
 
 def test_drop_inactive_elements():
@@ -464,39 +71,6 @@ def test_drop_inactive_elements():
     assert gen0 not in net.gen.index
 
 
-def test_get_connected_lines_at_bus():
-    net = pp.create_empty_network()
-
-    bus0 = pp.create_bus(net, 0.4)
-    bus1 = pp.create_bus(net, 0.4)
-
-    line0 = pp.create_line(net, bus0, bus1, length_km=1., std_type="NAYY 4x50 SE")
-    line1 = pp.create_line(net, bus0, bus1, length_km=1., std_type="NAYY 4x50 SE")
-    line2 = pp.create_line(net, bus0, bus1, in_service=False, length_km=1., std_type="NAYY 4x50 SE")
-    line3 = pp.create_line(net, bus0, bus1, length_km=1., std_type="NAYY 4x50 SE")
-
-    pp.create_switch(net, bus0, line0, "l")
-    pp.create_switch(net, bus0, line1, "l", closed=False)
-    pp.create_switch(net, bus0, line2, "l")
-
-    lines = pp.get_connected_elements(net, "line", bus0, respect_switches=False,
-                                      respect_in_service=False)
-
-    assert set(lines) == {line0, line1, line2, line3}
-
-    lines = pp.get_connected_elements(net, "line", bus0, respect_switches=True,
-                                      respect_in_service=False)
-    assert set(lines) == {line0, line2, line3}
-
-    lines = pp.get_connected_elements(net, "line", bus0, respect_switches=True,
-                                      respect_in_service=True)
-    assert set(lines) == {line0, line3}
-
-    lines = pp.get_connected_elements(net, "line", bus0, respect_switches=False,
-                                      respect_in_service=True)
-    assert set(lines) == {line0, line1, line3}
-
-
 def test_merge_indices():
     net1 = nw.create_cigre_network_mv()
     pp.create_pwl_cost(net1, 0, "load", [[0, 20, 1], [20, 30, 2]])
@@ -543,8 +117,8 @@ def test_merge_and_split_nets():
     pp.runpp(net)
 
     # check that merge_nets() doesn't change inputs (but result tables)
-    assert_net_equal(net1, net1_before, check_without_results=True)
-    assert_net_equal(net2, net2_before, check_without_results=True)
+    pp.test.assert_net_equal(net1, net1_before, check_without_results=True)
+    pp.test.assert_net_equal(net2, net2_before, check_without_results=True)
 
     # check that results of merge_nets() fit
     assert np.allclose(net.res_bus.vm_pu.iloc[:n1].values, net1.res_bus.vm_pu.values)
@@ -576,8 +150,8 @@ def test_merge_asymmetric():
     net2_before = copy.deepcopy(net2)
     net = pp.merge_nets(net1, net2, net2_reindex_log_level="debug")
 
-    assert_net_equal(net1, net1_before, check_without_results=True)
-    assert_net_equal(net2, net2_before, check_without_results=True)
+    pp.test.assert_net_equal(net1, net1_before, check_without_results=True)
+    pp.test.assert_net_equal(net2, net2_before, check_without_results=True)
     assert len(net.asymmetric_load.bus.unique()) == 2 * n_load_busses
     assert len(net.asymmetric_sgen.bus.unique()) == 2 * n_sgen_busses
 
@@ -669,48 +243,6 @@ def test_select_subnet():
     elements = set(net2.switch[net2.switch.et == 'b'].element)
     subnet = pp.select_subnet(net2, buses | elements)
     assert net2.switch[net2.switch.et == 'b'].index.isin(subnet.switch.index).all()
-
-
-def test_overloaded_lines():
-    net = pp.create_empty_network()
-
-    bus0 = pp.create_bus(net, vn_kv=.4)
-    bus1 = pp.create_bus(net, vn_kv=.4)
-
-    pp.create_ext_grid(net, bus0)
-
-    line0 = pp.create_line(net, bus0, bus1, length_km=1, std_type="NAYY 4x50 SE")
-    line1 = pp.create_line(net, bus0, bus1, length_km=1, std_type="NA2XS2Y 1x95 RM/25 12/20 kV")
-    line2 = pp.create_line(net, bus0, bus1, length_km=1, std_type="15-AL1/3-ST1A 0.4")
-    pp.create_line(net, bus0, bus1, length_km=10, std_type="149-AL1/24-ST1A 10.0")
-
-    pp.create_load(net, bus1, p_mw=0.2, q_mvar=0.05)
-
-    pp.runpp(net)
-    # test the overloaded lines by default value of max_load=100
-    overloaded_lines = pp.overloaded_lines(net, max_load=100)
-
-    assert set(overloaded_lines) == {line0, line1}
-
-    # test the overloaded lines by a self defined value of max_load=50
-    overloaded_lines = pp.overloaded_lines(net, max_load=50)
-
-    assert set(overloaded_lines) == {line0, line1, line2}
-
-
-def test_violated_buses():
-    net = nw.create_cigre_network_lv()
-
-    pp.runpp(net)
-
-    # set the range of vm.pu
-    min_vm_pu = 0.92
-    max_vm_pu = 1.1
-
-    # print out the list of violated_bus's index
-    violated_bus = pp.violated_buses(net, min_vm_pu, max_vm_pu)
-
-    assert set(violated_bus) == set(net["bus"].index[[16, 35, 36, 40]])
 
 
 def test_add_zones_to_elements():
@@ -826,58 +358,6 @@ def test_close_switch_at_line_with_two_open_switches():
 
     # assertion: sw2 closed
     assert net.switch.closed.loc[1]
-
-
-def test_pq_from_cosphi():
-    p, q = pp.pq_from_cosphi(1 / 0.95, 0.95, "underexcited", "load")
-    assert np.isclose(p, 1)
-    assert np.isclose(q, 0.3286841051788632)
-
-    s = np.array([1, 1, 1])
-    cosphi = np.array([1, 0.5, 0])
-    pmode = np.array(["load", "load", "load"])
-    qmode = np.array(["underexcited", "underexcited", "underexcited"])
-    p, q = pp.pq_from_cosphi(s, cosphi, qmode, pmode)
-    excpected_values = (np.array([1, 0.5, 0]), np.array([0, 0.8660254037844386, 1]))
-    assert np.allclose(p, excpected_values[0])
-    assert np.allclose(q, excpected_values[1])
-
-    pmode = "gen"
-    p, q = pp.pq_from_cosphi(s, cosphi, qmode, pmode)
-    assert np.allclose(p, excpected_values[0])
-    assert np.allclose(q, -excpected_values[1])
-
-    qmode = "overexcited"
-    p, q = pp.pq_from_cosphi(s, cosphi, qmode, pmode)
-    assert np.allclose(p, excpected_values[0])
-    assert np.allclose(q, excpected_values[1])
-
-    with pytest.raises(ValueError):
-        pp.pq_from_cosphi(1, 0.95, "ohm", "gen")
-
-    p, q = pp.pq_from_cosphi(0, 0.8, "overexcited", "gen")
-    assert np.isclose(p, 0)
-    assert np.isclose(q, 0)
-
-
-def test_cosphi_from_pq():
-    cosphi, s, qmode, pmode = pp.cosphi_from_pq(1, 0.4)
-    assert np.isclose(cosphi, 0.9284766908852593)
-    assert np.isclose(s, 1.077032961426901)
-    assert qmode == 'underexcited'
-    assert pmode == 'load'
-
-    p = np.array([1, 1, 1, 1, 1, 0, 0, 0, -1, -1, -1])
-    q = np.array([1, -1, 0, 0.5, -0.5, 1, -1, 0, 1, -1, 0])
-    cosphi, s, qmode, pmode = pp.cosphi_from_pq(p, q)
-    assert np.allclose(cosphi[[0, 1, 8, 9]], 2 ** 0.5 / 2)
-    assert np.allclose(cosphi[[3, 4]], 0.89442719)
-    assert np.allclose(cosphi[[2, 10]], 1)
-    assert pd.Series(cosphi[[5, 6, 7]]).isnull().all()
-    assert np.allclose(s, (p ** 2 + q ** 2) ** 0.5)
-    assert all(pmode == np.array(["load"] * 5 + ["undef"] * 3 + ["gen"] * 3))
-    ind_cap_ind = ["underexcited", "overexcited", "underexcited"]
-    assert all(qmode == np.array(ind_cap_ind + ["underexcited", "overexcited"] + ind_cap_ind * 2))
 
 
 def test_create_replacement_switch_for_branch():
@@ -1000,86 +480,6 @@ def test_all(net):
     assert net.impedance.in_service.at[0]
     assert ~net.impedance.in_service.at[1]
     assert net.impedance.in_service.at[2]
-
-
-def test_get_element_indices():
-    net = nw.example_multivoltage()
-    idx1 = pp.get_element_indices(net, "bus", ["Bus HV%i" % i for i in range(1, 4)])
-    idx2 = pp.get_element_indices(net, ["bus", "line"], "HV", exact_match=False)
-    idx3 = pp.get_element_indices(net, ["bus", "line"], ["Bus HV3", "MV Line6"])
-    assert [32, 33, 34] == idx1
-    assert ([32, 33, 34, 35] == idx2[0]).all()
-    assert ([0, 1, 2, 3, 4, 5] == idx2[1]).all()
-    assert [34, 11] == idx3
-
-
-def test_next_bus():
-    net = pp.create_empty_network()
-
-    bus0 = pp.create_bus(net, vn_kv=110)
-    bus1 = pp.create_bus(net, vn_kv=20)
-    bus2 = pp.create_bus(net, vn_kv=10)
-    bus3 = pp.create_bus(net, vn_kv=0.4)
-    bus4 = pp.create_bus(net, vn_kv=0.4)
-    bus5 = pp.create_bus(net, vn_kv=20)
-
-    trafo0 = pp.create_transformer3w(net, hv_bus=bus0, mv_bus=bus1, lv_bus=bus2, name='trafo0',
-                                     std_type='63/25/38 MVA 110/20/10 kV')
-    trafo1 = pp.create_transformer(net, hv_bus=bus2, lv_bus=bus3, std_type='0.4 MVA 10/0.4 kV')
-
-    line1 = pp.create_line(net, from_bus=bus3, to_bus=bus4, length_km=20.1,
-                           std_type='24-AL1/4-ST1A 0.4', name='line1')
-
-    # switch0=pp.create_switch(net, bus = bus0, element = trafo0, et = 't3') #~~~~~ not implementable now
-    switch1 = pp.create_switch(net, bus=bus1, element=bus5, et='b')
-    switch2 = pp.create_switch(net, bus=bus2, element=trafo1, et='t')
-    switch3 = pp.create_switch(net, bus=bus3, element=line1, et='l')
-
-    # assert pp.next_bus(net,bus0,trafo0,et='trafo3w')==bus1                         # not implemented in existing toolbox
-    # assert pp.next_bus(net,bus0,trafo0,et='trafo3w',choice_for_trafo3w='lv')==bus2 # not implemented in existing toolbox
-    assert pp.next_bus(net, bus1, switch1, et='switch') == bus5  # Switch with bus2bus connection
-    # assert not pp.next_bus(net,bus2,switch2,et='switch')==bus3  # Switch with bus2trafo connection:- gives trasformer id instead of bus id
-    assert pp.next_bus(net, bus2, trafo1, et='trafo') == bus3
-    # assert pp.next_bus(net,bus3,switch3,et='switch') ==bus4  # Switch with bus2line connection :- gives line id instead of bus id
-    assert pp.next_bus(net, bus3, line1, et='line') == bus4
-
-
-def test_get_connected_buses():
-    net = pp.create_empty_network()
-
-    bus0 = pp.create_bus(net, vn_kv=110)
-    bus1 = pp.create_bus(net, vn_kv=20)
-    bus2 = pp.create_bus(net, vn_kv=10)
-    bus3 = pp.create_bus(net, vn_kv=0.4)
-    bus4 = pp.create_bus(net, vn_kv=0.4)
-    bus5 = pp.create_bus(net, vn_kv=20)
-
-    trafo0 = pp.create_transformer3w(net, hv_bus=bus0, mv_bus=bus1, lv_bus=bus2,
-                                     std_type='63/25/38 MVA 110/20/10 kV')
-    trafo1 = pp.create_transformer(net, hv_bus=bus2, lv_bus=bus3, std_type='0.4 MVA 10/0.4 kV')
-    line1 = pp.create_line(net, from_bus=bus3, to_bus=bus4, length_km=20.1,
-                           std_type='24-AL1/4-ST1A 0.4')
-
-    switch0a = pp.create_switch(net, bus=bus0, element=trafo0, et='t3')
-    switch0b = pp.create_switch(net, bus=bus2, element=trafo0, et='t3')
-    switch1 = pp.create_switch(net, bus=bus1, element=bus5, et='b')
-    switch2 = pp.create_switch(net, bus=bus2, element=trafo1, et='t')
-    switch3 = pp.create_switch(net, bus=bus3, element=line1, et='l')
-
-    assert list(pp.get_connected_buses(net, [bus0])) == [bus1, bus2]
-    assert list(pp.get_connected_buses(net, [bus1])) == [bus0, bus2, bus5]
-    assert list(pp.get_connected_buses(net, [bus2])) == [bus0, bus1, bus3]
-    assert list(pp.get_connected_buses(net, [bus3])) == [bus2, bus4]
-    assert list(pp.get_connected_buses(net, [bus4])) == [bus3]
-    assert list(pp.get_connected_buses(net, [bus5])) == [bus1]
-    assert list(pp.get_connected_buses(net, [bus0, bus1])) == [bus2, bus5]
-    assert list(pp.get_connected_buses(net, [bus2, bus3])) == [bus0, bus1, bus4]
-
-    net.switch.loc[[switch0b, switch1, switch2, switch3], 'closed'] = False
-    assert list(pp.get_connected_buses(net, [bus0])) == [bus1]
-    assert list(pp.get_connected_buses(net, [bus1])) == [bus0]
-    assert list(pp.get_connected_buses(net, [bus3])) == []
-    assert list(pp.get_connected_buses(net, [bus4])) == []
 
 
 def test_drop_elements_at_buses():
@@ -1566,62 +966,6 @@ def test_merge_same_bus_generation_plants():
     idx_sgen22 = net.sgen.index[net.sgen.bus == 22]
     assert len(idx_sgen22) == 1
     assert np.isclose(net.sgen.q_mvar.at[idx_sgen22[0]], 20 + 21)  # correct value sum (q_mvar)
-
-
-def test_get_false_links():
-    net = pp.create_empty_network()
-    pp.create_buses(net, 6, 10, index=[0, 1, 3, 4, 6, 7])
-
-    # --- gens
-    pp.create_gens(net, [0, 1, 3], 5)
-    # manipulate to not existing
-    net.gen.bus.at[1] = 999
-
-    # --- sgens
-    pp.create_sgens(net, [0, 1, 3], 5)
-
-    # --- lines
-    for fbus, tbus in zip([0, 1, 4, 6, 7], [1, 4, 6, 7, 3]):
-        pp.create_line(net, fbus, tbus, 2., "NA2XS2Y 1x185 RM/25 6/10 kV")
-    # manipulate to not existing
-    net.line.from_bus.at[1] = 2
-    net.line.to_bus.at[4] = 999
-
-    # --- measurements
-    pp.create_measurement(net, "v", "bus", 1.01, 5, 1)
-    pp.create_measurement(net, "i", "line", 0.41, 1, 0, side="from")
-    pp.create_measurement(net, "i", "line", 0.41, 1, 2, side="from")
-    pp.create_measurement(net, "v", "bus", 1.01, 5, 6)
-    pp.create_measurement(net, "i", "line", 0.41, 1, 1, side="from")
-    # manipulate to not existing
-    net.measurement.element.at[1] = 999
-    net.measurement.element.at[3] = 999
-
-    # --- poly_cost
-    pp.create_poly_cost(net, 0, "gen", 5)
-    pp.create_poly_cost(net, 1, "gen", 5)
-    pp.create_poly_cost(net, 0, "sgen", 5)
-    pp.create_poly_cost(net, 1, "sgen", 5)
-    # manipulate to not existing
-    net.poly_cost.element.at[1] = 999
-    net.poly_cost.element.at[2] = 999
-
-    expected = {"gen": {1},
-                "line": {1, 4},
-                "measurement": {1, 3},
-                "poly_cost": {1, 2}}
-    determined = pp.false_elm_links_loop(net)
-    assert {elm: set(idx) for elm, idx in determined.items()} == expected
-
-
-def test_res_power_columns():
-    assert pp.res_power_columns("gen") == ["p_mw", "q_mvar"]
-    assert pp.res_power_columns("line") == pp.res_power_columns("line", side="from") == \
-        pp.res_power_columns("line", side=0) == ["p_from_mw", "q_from_mvar"]
-    assert pp.res_power_columns("line", side="all") == [
-        "p_from_mw", "q_from_mvar", "p_to_mw", "q_to_mvar"]
-    assert pp.res_power_columns("trafo3w", side="all") == [
-        "p_hv_mw", "q_hv_mvar", "p_mv_mw", "q_mv_mvar", "p_lv_mw", "q_lv_mvar"]
 
 
 if __name__ == '__main__':
