+++ pandapower/toolbox/general_issues.py
