+++ pandapower/toolbox/result_info.py
